renaming substrate files from Manager -> Implementation
